test: inplement the cucumber conformance tests from cloudevents/spec
This commit adds cucumber-js conformance steps and includes the cucumber
tests in 'npm test'.
