+++ test/integration/parser_test.ts
@@ -1,8 +1,8 @@
 import "mocha";
 import { expect } from "chai";
 
-import { JSONParser as Parser } from "../src/parsers/";
-import { ValidationError } from "../src/";
+import { JSONParser as Parser } from "../../src/parsers/";
+import { ValidationError } from "../../src/";
 
 describe("JSON Event Format Parser", () => {
   it("Throw error when payload is an integer", () => {
