+++ test/integration/spec_03_tests.ts
@@ -1,7 +1,7 @@
 import "mocha";
 import { expect } from "chai";
-import { CloudEvent, Version, ValidationError, Mode } from "../src";
-import Constants from "../src/constants";
+import { CloudEvent, Version, ValidationError, Mode } from "../../src";
+import Constants from "../../src/constants";
 
 const id = "97699ec2-a8d9-47c1-bfa0-ff7aa526f838";
 const type = "com.github.pull.create";
